Trimmed down the code for making powers clonable significantly. Added previously non-clonable powers to the clonable club. clonable.
