+++ mod/src/main/java/basemod/patches/com/megacrit/cardcrawl/powers/CloneablePowers/OwnerAmountCardPowerPatch.java
@@ -1,0 +1,18 @@
+package basemod.patches.com.megacrit.cardcrawl.powers.CloneablePowers;
+
+import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
+import com.megacrit.cardcrawl.powers.NightmarePower;
+import javassist.CannotCompileException;
+import javassist.CtBehavior;
+import javassist.NotFoundException;
+
+
+@SpirePatch(
+        clz = NightmarePower.class,
+        method = SpirePatch.CONSTRUCTOR
+)
+public class OwnerAmountCardPowerPatch extends CloneablePowersPatch {
+    public static void Raw(CtBehavior ctMethodToPatch) throws NotFoundException, CannotCompileException {
+        addMakeCopyMethod(ctMethodToPatch, "amount, card");
+    }
+}
