+++ mod/src/main/java/basemod/patches/com/megacrit/cardcrawl/powers/CloneablePowers/OwnerAmountNamePowerPatch.java
@@ -1,0 +1,24 @@
+package basemod.patches.com.megacrit.cardcrawl.powers.CloneablePowers;
+
+import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
+import com.megacrit.cardcrawl.core.AbstractCreature;
+import com.megacrit.cardcrawl.powers.NextTurnBlockPower;
+import javassist.CannotCompileException;
+import javassist.CtBehavior;
+import javassist.NotFoundException;
+
+
+@SpirePatch(
+        clz = NextTurnBlockPower.class,
+        method = SpirePatch.CONSTRUCTOR,
+        paramtypez = {
+                AbstractCreature.class,
+                int.class,
+                String.class
+        }
+)
+public class OwnerAmountNamePowerPatch extends CloneablePowersPatch {
+    public static void Raw(CtBehavior ctMethodToPatch) throws NotFoundException, CannotCompileException {
+        addMakeCopyMethod(ctMethodToPatch, "amount, name");
+    }
+}
